add ConsumerRecord timestamp extractor and set as default
